Refactor server router following Express conventions
