+++ server/src/main.ts
@@ -1,5 +1,9 @@
 #!/usr/bin/env node
 
+/**
+ * Script used to start a Disco Server
+ */
+
 import { defaultTasks } from "@epfml/discojs";
 
 import { Server } from "./server.js";
@@ -7,7 +11,7 @@ import { Server } from "./server.js";
 const PORT = 8080;
 
 const providers = Object.values(defaultTasks);
-
+// Init the server with default tasks
 const server = await Server.of(...providers);
 
 console.info("Server loaded the tasks below");
@@ -22,5 +26,5 @@ console.table(
     })),
 );
 
-await server.serve(PORT);
-console.log(`Disco Server listening on http://localhost:${PORT}`);
+const [_, serverURL] = await server.serve(PORT);
+console.log(`Disco Server listening on ${serverURL.toString()}`);
