--- server/src/router/decentralized/index.ts
@@ -1,1 +1,0 @@
-export { Decentralized } from './server.js'
