--- server/src/router/federated/index.ts
@@ -1,1 +1,0 @@
-export { Federated } from './server.js'
