--- server/src/router/index.ts
@@ -1,1 +1,0 @@
-export { Router } from './router.js'
