+++ server/src/routes/index.ts
@@ -1,0 +1,2 @@
+export { TaskRouter } from "./tasks.js";
+export { FederatedRouter, DecentralizedRouter } from "./training/index.js";
