+++ server/src/routes/training/index.ts
@@ -1,0 +1,2 @@
+export { FederatedRouter } from "./federated.js";
+export { DecentralizedRouter } from "./decentralized.js";
