+++ server/src/server.ts
@@ -1,3 +1,4 @@
+import createDebug from "debug";
 import cors from "cors";
 import express from "express";
 import expressWS from "express-ws";
@@ -5,12 +6,22 @@ import type * as http from "http";
 
 import type { TaskProvider } from "@epfml/discojs";
 
-import { Router } from "./router/index.js";
+import { TaskRouter, FederatedRouter, DecentralizedRouter } from './routes/index.js'
 import { TasksAndModels } from "./tasks.js";
 
+const debug = createDebug("server");
+
+/**
+ * The Disco Server, a wrapper of the Express app
+ * Its main goal is to provide the available tasks (DISCOllaboratives)
+ * and tasks' base models to clients. 
+ * New tasks can be added via the `addTask` method.
+ * Most of the logic is deferred to the Router abstraction. 
+ */
 export class Server {
   readonly #tasksAndModels = new TasksAndModels();
 
+  // Static method to asynchronously init the Server
   static async of(...tasks: TaskProvider[]): Promise<Server> {
     const ret = new Server();
     await Promise.all(tasks.map((t) => ret.addTask(t)));
@@ -37,8 +48,27 @@ export class Server {
     app.use(express.json({ limit: "50mb" }));
     app.use(express.urlencoded({ limit: "50mb", extended: false }));
 
-    const baseRouter = new Router(wsApplier, this.#tasksAndModels);
-    app.use("/", baseRouter.router);
+    const taskRouter = new TaskRouter(this.#tasksAndModels)
+    const federatedRouter = new FederatedRouter(wsApplier, this.#tasksAndModels)
+    const decentralizedRouter = new DecentralizedRouter(wsApplier, this.#tasksAndModels)
+
+    process.nextTick(() =>
+      wsApplier.getWss().on('connection', (ws, req) => {
+        if (!federatedRouter.isValidUrl(req.url) && !decentralizedRouter.isValidUrl(req.url)) {
+          debug("connection refused on %s", req.url);
+          ws.terminate()
+          ws.close()
+        }
+      })
+    )
+
+    app.get('/', (_, res, next) => {
+      res.send('The DISCO Server\n')
+      next()
+    })
+    app.use('/feai', federatedRouter.router)
+    app.use('/deai', decentralizedRouter.router)
+    app.use('/tasks', taskRouter.router)
 
     const server = await new Promise<http.Server>((resolve, reject) => {
       const ret = app.listen(port);
